Revision de constraints, implementacion de relaciones

+++ src/main/java/org/springframework/gresur/repository/NotificacionRepository.java
@@ -3,6 +3,6 @@ package org.springframework.gresur.repository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.gresur.model.Notificacion;
 
-public interface NotificacionRepository extends CrudRepository<Notificacion, Integer>{
+public interface NotificacionRepository extends CrudRepository<Notificacion, Long>{
 
 }
